refactor: add business logic separation for the BookTable component
