+++ src/context/bookingData/BookingDataContext.tsx
@@ -1,5 +1,4 @@
 import { createContext } from 'react'
-import { Props } from './BookingDataContext.types'
-import { INITIAL_BOOKING_STATE } from '../../constants/booking'
+import { BookingData } from './BookingDataContext.types'
 
-export const BookingDataContext = createContext<Props | undefined>(undefined)
+export const BookingDataContext = createContext<BookingData | undefined>(undefined!)
