+++ src/context/bookingData/BookingDataContext.types.ts
@@ -1,6 +1,6 @@
 import { INITIAL_BOOKING_STATE_TYPE } from "../../constants/booking";
 
-export type Props = {
+export type BookingData = {
   booking: Booking
   setBooking?: React.Dispatch<React.SetStateAction<INITIAL_BOOKING_STATE_TYPE>>
   handleDateChange?: (date: Date | null, e: React.SyntheticEvent<any, Event> | undefined) => void
